requirements.txt
acasearch==0.1
arxiv==0.5.1
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
colorlog==4.1.0
dj-database-url==0.5.0
Django==3.0.2
django-heroku==0.3.1
djangorestframework==3.11.0
feedparser==5.2.1
gunicorn==20.0.4
idna==2.8
psycopg2==2.8.4
pytest-runner==5.2
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
